Rename sync to UniversalRouterSync
