+++ tools/build.js
@@ -31,34 +31,6 @@ const files = [
     external: ['path-to-regexp'],
   },
   {
-    input: 'dist/src/generateUrls.js',
-    output: 'dist/generateUrls/index.js',
-    format: 'cjs',
-    external: ['path-to-regexp', path.resolve('dist/src/UniversalRouter.js')],
-    paths: { [path.resolve('dist/src/UniversalRouter.js')]: '..' },
-  },
-  {
-    input: 'dist/src/generateUrls.js',
-    output: 'dist/generateUrls/module.js',
-    format: 'es',
-    external: ['path-to-regexp', path.resolve('dist/src/UniversalRouter.js')],
-    paths: { [path.resolve('dist/src/UniversalRouter.js')]: '..' },
-  },
-  {
-    input: 'dist/src/sync.js',
-    output: 'dist/sync/index.js',
-    format: 'cjs',
-    external: ['path-to-regexp', path.resolve('dist/src/UniversalRouter.js')],
-    paths: { [path.resolve('dist/src/UniversalRouter.js')]: '..' },
-  },
-  {
-    input: 'dist/src/sync.js',
-    output: 'dist/sync/module.js',
-    format: 'es',
-    external: ['path-to-regexp', path.resolve('dist/src/UniversalRouter.js')],
-    paths: { [path.resolve('dist/src/UniversalRouter.js')]: '..' },
-  },
-  {
     input: 'dist/src/UniversalRouter.js',
     output: 'dist/universal-router.js',
     format: 'umd',
@@ -71,6 +43,56 @@ const files = [
     format: 'umd',
     name: 'UniversalRouter',
     external: [],
+  },
+
+  // sync add-on
+  {
+    input: 'dist/src/UniversalRouterSync.js',
+    output: 'dist/sync/index.js',
+    format: 'cjs',
+    external: ['path-to-regexp', path.resolve('dist/src/UniversalRouter.js')],
+    paths: { [path.resolve('dist/src/UniversalRouter.js')]: '..' },
+  },
+  {
+    input: 'dist/src/UniversalRouterSync.js',
+    output: 'dist/sync/module.js',
+    format: 'es',
+    external: ['path-to-regexp', path.resolve('dist/src/UniversalRouter.js')],
+    paths: { [path.resolve('dist/src/UniversalRouter.js')]: '..' },
+  },
+  {
+    input: 'dist/src/UniversalRouterSync.js',
+    output: 'dist/universal-router-sync.js',
+    format: 'umd',
+    name: 'sync',
+    external: ['path-to-regexp', path.resolve('dist/src/UniversalRouter.js')],
+    paths: { [path.resolve('dist/src/UniversalRouter.js')]: './universal-router.js' },
+    globals: { [path.resolve('dist/src/UniversalRouter.js')]: 'UniversalRouter' },
+  },
+  {
+    input: 'dist/src/UniversalRouterSync.js',
+    output: 'dist/universal-router-sync.min.js',
+    format: 'umd',
+    name: 'sync',
+    external: ['path-to-regexp', path.resolve('dist/src/UniversalRouter.js')],
+    paths: { [path.resolve('dist/src/UniversalRouter.js')]: './universal-router.min.js' },
+    globals: { [path.resolve('dist/src/UniversalRouter.js')]: 'UniversalRouter' },
+  },
+
+  // generateUrls add-on
+  {
+    input: 'dist/src/generateUrls.js',
+    output: 'dist/generateUrls/index.js',
+    format: 'cjs',
+    external: ['path-to-regexp', path.resolve('dist/src/UniversalRouter.js')],
+    paths: { [path.resolve('dist/src/UniversalRouter.js')]: '..' },
+  },
+  {
+    input: 'dist/src/generateUrls.js',
+    output: 'dist/generateUrls/module.js',
+    format: 'es',
+    external: ['path-to-regexp', path.resolve('dist/src/UniversalRouter.js')],
+    paths: { [path.resolve('dist/src/UniversalRouter.js')]: '..' },
   },
   {
     input: 'dist/src/generateUrls.js',
@@ -86,24 +108,6 @@ const files = [
     output: 'dist/universal-router-generate-urls.min.js',
     format: 'umd',
     name: 'generateUrls',
-    external: ['path-to-regexp', path.resolve('dist/src/UniversalRouter.js')],
-    paths: { [path.resolve('dist/src/UniversalRouter.js')]: './universal-router.min.js' },
-    globals: { [path.resolve('dist/src/UniversalRouter.js')]: 'UniversalRouter' },
-  },
-  {
-    input: 'dist/src/sync.js',
-    output: 'dist/universal-router-sync.js',
-    format: 'umd',
-    name: 'sync',
-    external: ['path-to-regexp', path.resolve('dist/src/UniversalRouter.js')],
-    paths: { [path.resolve('dist/src/UniversalRouter.js')]: './universal-router.js' },
-    globals: { [path.resolve('dist/src/UniversalRouter.js')]: 'UniversalRouter' },
-  },
-  {
-    input: 'dist/src/sync.js',
-    output: 'dist/universal-router-sync.min.js',
-    format: 'umd',
-    name: 'sync',
     external: ['path-to-regexp', path.resolve('dist/src/UniversalRouter.js')],
     paths: { [path.resolve('dist/src/UniversalRouter.js')]: './universal-router.min.js' },
     globals: { [path.resolve('dist/src/UniversalRouter.js')]: 'UniversalRouter' },
@@ -168,6 +172,19 @@ async function build() {
   delete libPkg.scripts
   await fs.outputJson('dist/package.json', libPkg, { spaces: 2 })
 
+  // Create sync/package.json for convenient import
+  const syncPkg = {
+    ...pkg,
+    name: 'sync',
+    description: 'Universal Router Sync Add-on',
+    esnext: '../src/sync.js',
+    types: '../src/sync.d.ts',
+  }
+  delete syncPkg.dependencies
+  delete syncPkg.devDependencies
+  delete syncPkg.scripts
+  await fs.outputJson('dist/sync/package.json', syncPkg, { spaces: 2 })
+
   // Create generateUrls/package.json for convenient import
   const generateUrlsPkg = {
     ...pkg,
@@ -180,19 +197,6 @@ async function build() {
   delete generateUrlsPkg.devDependencies
   delete generateUrlsPkg.scripts
   await fs.outputJson('dist/generateUrls/package.json', generateUrlsPkg, { spaces: 2 })
-
-  // Create sync/package.json for convenient import
-  const syncPkg = {
-    ...pkg,
-    name: 'sync',
-    description: 'Universal Router Generate URLs Add-on',
-    esnext: '../src/sync.js',
-    types: '../src/sync.d.ts',
-  }
-  delete syncPkg.dependencies
-  delete syncPkg.devDependencies
-  delete syncPkg.scripts
-  await fs.outputJson('dist/sync/package.json', syncPkg, { spaces: 2 })
 }
 
 module.exports = build()
